requirements.txt
numpy
streamlit
pillow
pydicom
numba

